Refactor My Store package
